Refactored place_patrol_boat_player_one into patrol_boat.py.
