+++ test_patrol_boat.py
@@ -24,3 +24,17 @@ class TestPatrol_Boat(unittest.TestCase):
 
         # assert
         self.assertEqual(expected_result, observed_result)
+
+    @patch('ship.Ship.get_hit_counter_player', return_value=[2, 3, 1, 2, 0])
+    def test_ship_sunk_patrol_boat_player(self, get_hit_counter_player):
+        # create an object of class Patrol_Boat
+        self.patrol_boat = Patrol_Boat()
+        self.constants = utils.Constants()
+
+        # when
+        expected_result = True
+
+        # call method of class BattleShip
+        observed_result = self.patrol_boat.ship_sunk_patrol_boat_player()
+
+        self.assertEqual(expected_result, observed_result)
